% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_test.R
\name{conf_test}
\alias{conf_test}
\title{Conformal test for conditional distributions}
\usage{
conf_test(
  X1,
  X2,
  Y1,
  Y2,
  prop = 0.5,
  folds = NULL,
  bandwidth = NULL,
  est_method = c("KLR", "LL", "QL", "NN", "KLIEP")
)
}
\arguments{
\item{X1}{A matrix or data frame of covariates for the first sample.}

\item{X2}{A matrix or data frame of covariates for the second sample.}

\item{Y1}{A matrix or data frame of outcomes for the first sample.}

\item{Y2}{A matrix or data frame of outcomes for the second sample.}

\item{prop}{The proportion of data to be used for calculating the test statistic.
Default is 0.5.}

\item{folds}{The number of folds for cross-validation. Default is NULL.}

\item{bandwidth}{The bandwidth for kernel density estimation. Default is NULL.}

\item{est_method}{The method for estimating the density ratio.
One of "KLR", "LL", "QL", "NN", "KLIEP". Default is "KLR".}
}
\value{
A list containing the test statistic, variance, z-score, and p-value.
}
\description{
This function implements the two-sample conditional distribution
test using conformal prediction, proposed by Hu and Lei (2024).
}
\references{
Hu, X., & Lei, J. (2024). A two-sample conditional distribution
test using conformal prediction and weighted rank sum. Journal of the American
Statistical Association, 119(546), 1136-1154.
}
