% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_objective}
\alias{cv_objective}
\title{CV Objective Function (C++ Backend)}
\usage{
cv_objective(
  C,
  X,
  kY,
  local_weights,
  kappa,
  nu,
  kern_type,
  lower_C,
  tiny_num = 1e-12
)
}
\arguments{
\item{C}{A numeric vector of scaling factors for the bandwidth.}

\item{X}{A numeric matrix of predictors.}

\item{kY}{The precomputed RKHS kernel matrix for the response.}

\item{local_weights}{A numeric vector of pre-computed weights for each
observation, used for locally weighted CV. For standard CV, this should
be a vector of ones.}

\item{kappa}{The exponent for the sample size in the bandwidth formula.}

\item{nu}{The order of the smoothing kernel.}

\item{kern_type}{An integer mapping to the kernel function.}

\item{lower_C}{The lower bound for C values.}

\item{tiny_num}{A small number to prevent division by zero.}
}
\description{
Calculates the (potentially locally weighted) leave-one-out
cross-validation loss.
}
\keyword{internal}
