% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecf_test.R
\name{ecf_test}
\alias{ecf_test}
\title{Test for the equality of two regression curves via empirical characteristic functions}
\usage{
ecf_test(
  x1,
  x2,
  y1,
  y2,
  B = 299,
  h = NULL,
  C = NULL,
  rescale = FALSE,
  kern_smooth = c("epanechnikov", "gaussian", "uniform"),
  reg_method = c("llinear", "lconst")
)
}
\arguments{
\item{x1}{A numeric vector of the covariate for the first sample.}

\item{x2}{A numeric vector of the covariate for the second sample.}

\item{y1}{A numeric vector of the outcome for the first sample.}

\item{y2}{A numeric vector of the outcome for the second sample.}

\item{B}{The number of bootstrap replications to compute the p-value. Default is 299.}

\item{h}{The bandwidth parameter for the kernel smoothing.
If NULL (the default), it is selected automatically.}

\item{C}{A constant to multiply the automatically selected bandwidth. Default is 1.}

\item{rescale}{A logical indicating whether to rescale the covariate \code{x} to
the interval \verb{[0, 1]}.
Default is FALSE.}

\item{kern_smooth}{The kernel function to be used for smoothing. One of
"epanechnikov", "gaussian", or "uniform".
Default is "epanechnikov".}

\item{reg_method}{The regression method to be used.
One of "llinear" (local linear) or "lconst" (local constant, Nadaraya-Watson).
Default is "llinear".}
}
\value{
A list with the following components:
\item{stat}{The value of the test statistic.}
\item{pvalue}{The bootstrap p-value.}
\item{h}{The bandwidth used.}
\item{Tn_b}{The bootstrap values of the test statistic.}
}
\description{
Implements the test for the equality of two regression curves
proposed by Pardo-Fernandez et al. (2015).
The test is based on the difference between the empirical characteristic
functions of the residuals in the two groups.
}
\references{
Pardo‐Fernández, J. C., Jiménez‐Gamero, M. D., & Ghouch, A. E. (2015).
A Non‐parametric anova‐type Test for Regression Curves Based on Characteristic Functions.
Scandinavian Journal of Statistics, 42(1), 197-213.
}
