% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCDT-package.R
\docType{package}
\name{TCDT-package}
\alias{TCDT}
\alias{TCDT-package}
\title{TCDT: Two-Sample Conditional Distribution Testing}
\description{
This package provides tools for performing two-sample tests to determine if two conditional distributions are the same. It implements the methods proposed by Yan, Li and Zhang (2025), using either conditional energy distance (CED) or conditional maximum mean discrepancy (CMMD).
}
\author{
\strong{Maintainer}: Zhuoxi Li \email{lizhuoxi@ust.hk}

Authors:
\itemize{
  \item Jian Yan \email{jy929@cornell.edu}
  \item Xianyang Zhang \email{zhangxiany@stat.tamu.edu}
}

}
\keyword{internal}
